inline Root.render into child part update
